+++ src/py_dss_toolkit/results/SnapShot/SnapShotPowerFlowResults.py
@@ -4,8 +4,10 @@
 
 from py_dss_interface import DSS
 
+from py_dss_toolkit.results.SnapShot.AllLosses import AllLosses
 from py_dss_toolkit.results.SnapShot.CircuitSnapShotPowerFlowResults import CircuitSnapShotPowerFlowResults
 from py_dss_toolkit.results.SnapShot.Currents import Currents
+from py_dss_toolkit.results.SnapShot.Losses import Losses
 from py_dss_toolkit.results.SnapShot.Powers import Powers
 from py_dss_toolkit.results.SnapShot.VoltagesElement import VoltagesElement
 from py_dss_toolkit.results.SnapShot.VoltagesNodal import VoltagesNodal
@@ -18,6 +20,8 @@ class SnapShotPowerFlowResults(VoltagesNodal,
                                VoltagesElement,
                                Currents,
                                Powers,
+                               Losses,
+                               AllLosses,
                                CircuitSnapShotPowerFlowResults,
                                VoltagesNodalViolations,
                                CurrentsViolations,
@@ -28,6 +32,8 @@ class SnapShotPowerFlowResults(VoltagesNodal,
         VoltagesElement.__init__(self, self._dss)
         Currents.__init__(self, self._dss)
         Powers.__init__(self, self._dss)
+        Losses.__init__(self, self._dss)
+        AllLosses.__init__(self, self._dss)
         CircuitSnapShotPowerFlowResults.__init__(self, self._dss)
         VoltagesNodalViolations.__init__(self, self._dss)
         CurrentsViolations.__init__(self, self._dss)